Revert "Fix an import issue in fbcode. (#10604)"
This reverts commit 6cd8133035fe06be4f629b69cfd0048e07709a84.
