+++ TARGETS
@@ -253,7 +253,6 @@ cpp_library_wrapper(name="rocksdb_lib", srcs=[
         "util/ribbon_config.cc",
         "util/slice.cc",
         "util/status.cc",
-        "util/stderr_logger.cc",
         "util/string_util.cc",
         "util/thread_local.cc",
         "util/threadpool_imp.cc",
@@ -592,7 +591,6 @@ cpp_library_wrapper(name="rocksdb_whole_archive_lib", srcs=[
         "util/ribbon_config.cc",
         "util/slice.cc",
         "util/status.cc",
-        "util/stderr_logger.cc",
         "util/string_util.cc",
         "util/thread_local.cc",
         "util/threadpool_imp.cc",
